Convert API files to TS
