+++ apps/prairielearn/src/api/v1/endpoints/courseInstanceAssessments/index.ts
@@ -1,4 +1,3 @@
-// @ts-check
 import * as path from 'node:path';
 
 import { Router } from 'express';
