+++ apps/prairielearn/src/api/v1/endpoints/courseInstanceGradebook/index.ts
@@ -1,18 +1,15 @@
-// @ts-check
-import * as path from 'node:path';
-
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
 
 import * as sqldb from '@prairielearn/postgres';
 
-const sql = sqldb.loadSql(path.join(import.meta.dirname, '..', 'queries.sql'));
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 const router = Router({ mergeParams: true });
 
 router.get(
   '/',
   asyncHandler(async (req, res) => {
-    const result = await sqldb.queryOneRowAsync(sql.select_course_instance_info, {
+    const result = await sqldb.queryOneRowAsync(sql.select_user_scores, {
       course_instance_id: res.locals.course_instance.id,
     });
     res.status(200).send(result.rows[0].item);
