--- apps/prairielearn/src/api/v1/notFound.js
@@ -1,7 +1,0 @@
-// @ts-check
-
-export default function (req, res, _next) {
-  res.status(404).send({
-    message: 'Not Found',
-  });
-}
