+++ apps/prairielearn/src/api/v1/notFound.ts
@@ -1,0 +1,7 @@
+import type { Request, Response } from 'express';
+
+export default function (req: Request, res: Response) {
+  res.status(404).send({
+    message: 'Not Found',
+  });
+}
