+++ apps/prairielearn/src/api/v1/prettyPrintJson.ts
@@ -1,11 +1,12 @@
-// @ts-check
+import { NextFunction, Request, Response } from 'express';
 
-export default function (req, res, next) {
+export default function (req: Request, res: Response, next: NextFunction) {
   res.json = (body) => {
     if (!res.get('Content-Type')) {
       res.set('Content-Type', 'application/json');
     }
     res.send(JSON.stringify(body, null, 2));
+    return res;
   };
   next();
 }
